@@ -3,10 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/react": "^1.6.7",
-    "@craco/craco": "^6.1.2",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
+    "@nextui-org/react": "^1.0.5-beta.5",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
@@ -19,19 +18,17 @@
     "@types/redux": "^3.6.0",
     "aos": "^2.3.4",
     "axios": "^0.21.1",
-    "dotenv": "^8.2.0",
     "framer-motion": "^4",
     "hamburger-react": "^2.4.1",
     "i18next": "^20.3.1",
     "i18next-http-backend": "^1.2.6",
-    "node-sass": "^4.14.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-i18next": "^11.11.0",
     "react-icons": "^4.2.0",
     "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.0",
     "redux": "^4.1.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-saga": "^1.1.3",
@@ -41,9 +38,9 @@
     "web-vitals": "^1.0.1"
   },
   "scripts": {
-    "build": "craco build",
-    "start": "SET PORT=3000 && craco start -p $PORT",
-    "test": "craco test",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -65,10 +62,10 @@
     ]
   },
   "devDependencies": {
-    "@types/react-router-hash-link": "^2.4.0",
     "@types/styled-components": "^5.1.11",
     "autoprefixer": "^9.8.6",
     "postcss": "^7.0.36",
+    "sass": "^1.49.10",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.1.4"
   }
 }
